requirements.txt
absl-py==2.1.0
accelerate==0.30.1
aiohttp==3.9.4
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
av==11.0.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
black==24.4.2
bleach==6.1.0
blinker==1.7.0
build==1.2.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.15
click==8.1.7
coloredlogs==15.0.1
comm==0.2.2
ctranslate2==4.2.1
dataclasses-json==0.6.4
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
docopt==0.6.2
easyocr==1.7.1
executing==2.0.1
faiss-cpu==1.8.0
fastapi==0.110.1
faster-whisper==1.0.1
fastjsonschema==2.19.1
filelock==3.13.4
flatbuffers==24.3.25
freetype-py==2.4.0
frozenlist==1.4.1
fsspec==2024.3.1
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
glob2==0.7
google-auth==2.29.0
google-pasta==0.2.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.1
h11==0.14.0
h5py==3.11.0
hsluv==5.0.4
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.0
humanfriendly==10.0
idna==3.7
imageio==2.34.1
importlib-metadata==7.0.0
importlib_resources==6.4.0
ipykernel==6.29.4
ipython==8.12.3
jedi==0.19.1
jellyfish==1.0.3
Jinja2==3.1.3
joblib==1.4.0
jsonpatch==1.33
jsonpickle==3.0.4
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keras==3.2.1
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.1.16
langchain-chroma==0.1.0
langchain-community==0.0.36
langchain-core==0.1.50
langchain-openai==0.1.6
langchain-text-splitters==0.0.1
langsmith==0.1.47
lazy_loader==0.4
libclang==18.1.1
llama-index==0.10.36
llama-index-agent-openai==0.2.4
llama-index-cli==0.1.12
llama-index-core==0.10.36
llama-index-embeddings-huggingface==0.2.0
llama-index-embeddings-langchain==0.1.2
llama-index-embeddings-openai==0.1.9
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-anyscale==0.1.4
llama-index-llms-huggingface==0.2.0
llama-index-llms-langchain==0.1.3
llama-index-llms-openai==0.1.18
llama-index-multi-modal-llms-openai==0.1.5
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.22
llama-index-readers-llama-parse==0.1.4
llama-parse==0.4.2
llamaindex-py-client==0.1.19
llvmlite==0.42.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib-inline==0.1.6
mdurl==0.1.2
minijinja==2.0.1
mistune==3.0.2
ml-dtypes==0.3.2
mmh3==4.1.0
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
namex==0.0.7
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
nltk==3.8.1
numba==0.59.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.3
openai==1.25.1
opencv-python-headless==4.9.0.80
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.0
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
patsy==0.5.6
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pipreqs==0.5.0
platformdirs==4.2.0
plotly==5.20.0
plotly-express==0.4.1
posthog==3.5.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pulsar-client==3.5.0
pure-eval==0.2.2
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pyclipper==1.3.0.post5
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
pydeck==0.8.1b0
Pygments==2.17.2
PyMuPDF==1.24.1
PyMuPDFb==1.24.1
pypdf==4.2.0
PyPika==0.48.9
pyproject_hooks==1.0.0
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyvis==0.3.2
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
sacremoses==0.1.1
safetensors==0.4.2
scikit-image==0.23.2
scikit-learn==1.4.2
scipy==1.13.0
sentence-transformers==2.6.1
sentencepiece==0.2.0
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.5
SQLAlchemy==2.0.29
stack-data==0.6.3
starlette==0.37.2
statsmodels==0.14.1
streamlit==1.34.0
striprtf==0.0.26
sympy==1.12
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
text-generation==0.7.0
threadpoolctl==3.4.0
tifffile==2024.5.10
tiktoken==0.6.0
tinycss2==1.3.0
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
transformers==4.39.3
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
vispy==0.14.2
watchdog==4.0.0
watchfiles==0.21.0
Wave==0.0.2
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.2
wrapt==1.16.0
yarg==0.1.9
yarl==1.9.4
zipp==3.18.1
audio_recorder_streamlit